add _JAVA_OPTIONS environment variable to avoid Maven error
This change is needed to use the surefire plugin on Maven which is used by many Java projects.
This is caused because of the version of OpenJDK used on the Debian Stretch Docker image which is then used by Jenkins.
This Java Option ignores the error and proceeds. This is the only viable workaround without having to add an option to all projects.